requirements.txt
annotated-types==0.7.0
anyio==4.4.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.7.4
chardet==5.2.0
distro==1.9.0
docx2txt==0.8
fulltext==0.5
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.8
iniconfig==2.0.0
lxml==5.3.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
python-dotenv==1.0.1
python-magic==0.4.27
python-pptx==1.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
xlrd==2.0.1
XlsxWriter==3.2.0
